requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bcrypt==5.0.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
ecdsa==0.19.1
fastapi==0.124.4
greenlet==3.3.0
gunicorn==23.0.0
h11==0.16.0
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==25.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
rich==14.2.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
